requirements.txt
bleak==0.14.3
paho-mqtt==1.6.1